Trianglegrid

Takes a bounding box and a cell depth and returns a set of Polygon's in a grid.

Usage

geo_trianglegrid(bbox, cell_size, units = "kilometers")

Arguments

bbox

extent in [minX, minY, maxX, maxY] order

cell_size

(numeric) the distance across each cell

units

(character) used in calculating cellSize, can be degrees, radians, miles, or kilometers (default)

Value

FeatureCollection<Polygon> grid of polygons

Examples

geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 3, "miles")
#> [1] "{\"features\":[{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.763205893570515],[-77.3876,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.763205893570515],[-77.3876,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.763205893570515],[-77.3876,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.763205893570515],[-77.3876,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.763205893570515],[-77.3319666495061,38.763205893570515],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.7198],[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.7198],[-77.3319666495061,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3319666495061,38.763205893570515],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.7198],[-77.3319666495061,38.763205893570515],[-77.3319666495061,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.763205893570515],[-77.2763332990122,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.7198],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.2763332990122,38.7198],[-77.2763332990122,38.763205893570515],[-77.22069994851829,38.763205893570515],[-77.2763332990122,38.7198],[-77.2763332990122,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.7198],[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.7198],[-77.22069994851829,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.22069994851829,38.763205893570515],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.7198],[-77.22069994851829,38.763205893570515],[-77.22069994851829,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.763205893570515],[-77.16506659802438,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.7198],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.16506659802438,38.7198],[-77.16506659802438,38.763205893570515],[-77.10943324753048,38.763205893570515],[-77.16506659802438,38.7198],[-77.16506659802438,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.7198],[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.7198],[-77.10943324753048,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.10943324753048,38.763205893570515],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.7198],[-77.10943324753048,38.763205893570515],[-77.10943324753048,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.763205893570515],[-77.05379989703657,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.7198],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.05379989703657,38.7198],[-77.05379989703657,38.763205893570515],[-76.99816654654266,38.763205893570515],[-77.05379989703657,38.7198],[-77.05379989703657,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.7198],[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.7198],[-76.99816654654266,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-76.99816654654266,38.763205893570515],[-76.94253319604876,38.763205893570515],[-76.94253319604876,38.7198],[-76.99816654654266,38.763205893570515],[-76.99816654654266,38.763205893570515]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"}],\"type\":\"FeatureCollection\"}"
geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 10, "miles")
#> [1] "{\"features\":[{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.86448631190172],[-77.20215549835363,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.86448631190172],[-77.20215549835363,38.86448631190172],[-77.20215549835363,38.7198],[-77.3876,38.86448631190172],[-77.3876,38.86448631190172]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.20215549835363,38.86448631190172],[-77.20215549835363,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.86448631190172],[-77.20215549835363,38.86448631190172],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,38.86448631190172],[-77.20215549835363,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,38.86448631190172],[-77.20215549835363,38.86448631190172],[-77.20215549835363,38.7198],[-77.3876,38.86448631190172],[-77.3876,38.86448631190172]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.7198],[-77.20215549835363,38.86448631190172],[-77.01671099670726,38.86448631190172],[-77.20215549835363,38.7198],[-77.20215549835363,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.7198],[-77.01671099670726,38.86448631190172],[-77.01671099670726,38.7198],[-77.20215549835363,38.7198],[-77.20215549835363,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.7198],[-77.20215549835363,38.86448631190172],[-77.01671099670726,38.7198],[-77.20215549835363,38.7198],[-77.20215549835363,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.86448631190172],[-77.01671099670726,38.86448631190172],[-77.01671099670726,38.7198],[-77.20215549835363,38.86448631190172],[-77.20215549835363,38.86448631190172]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.7198],[-77.20215549835363,38.86448631190172],[-77.01671099670726,38.86448631190172],[-77.20215549835363,38.7198],[-77.20215549835363,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.20215549835363,38.7198],[-77.01671099670726,38.86448631190172],[-77.01671099670726,38.7198],[-77.20215549835363,38.7198],[-77.20215549835363,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.7198],[-77.01671099670726,38.86448631190172],[-76.83126649506089,38.7198],[-77.01671099670726,38.7198],[-77.01671099670726,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.86448631190172],[-76.83126649506089,38.86448631190172],[-76.83126649506089,38.7198],[-77.01671099670726,38.86448631190172],[-77.01671099670726,38.86448631190172]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.7198],[-76.83126649506089,38.86448631190172],[-76.83126649506089,38.7198],[-77.01671099670726,38.7198],[-77.01671099670726,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.7198],[-77.01671099670726,38.86448631190172],[-76.83126649506089,38.86448631190172],[-77.01671099670726,38.7198],[-77.01671099670726,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.7198],[-77.01671099670726,38.86448631190172],[-76.83126649506089,38.7198],[-77.01671099670726,38.7198],[-77.01671099670726,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.01671099670726,38.86448631190172],[-76.83126649506089,38.86448631190172],[-76.83126649506089,38.7198],[-77.01671099670726,38.86448631190172],[-77.01671099670726,38.86448631190172]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"}],\"type\":\"FeatureCollection\"}"
geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, "miles")
#> [1] "{\"features\":[{\"geometry\":{\"coordinates\":[[[-77.3876,38.7198],[-77.3876,39.153858935705166],[-76.8312664950609,38.7198],[-77.3876,38.7198],[-77.3876,38.7198]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"},{\"geometry\":{\"coordinates\":[[[-77.3876,39.153858935705166],[-76.8312664950609,39.153858935705166],[-76.8312664950609,38.7198],[-77.3876,39.153858935705166],[-77.3876,39.153858935705166]]],\"type\":\"Polygon\"},\"properties\":{},\"type\":\"Feature\"}],\"type\":\"FeatureCollection\"}"